feat: astrbot external api
