+++ astrbot/dashboard/entities.py
@@ -1,0 +1,21 @@
+from pydantic.dataclasses import dataclass
+
+
+@dataclass
+class Response:
+    status: str | None = None
+    message: str | None = None
+    data: dict | list | None = None
+
+    def error(self, message: str):
+        self.status = "error"
+        self.message = message
+        return self
+
+    def ok(self, data: dict | list | None = None, message: str | None = None):
+        self.status = "ok"
+        if data is None:
+            data = {}
+        self.data = data
+        self.message = message
+        return self
