+++ astrbot/dashboard/routes/route.py
@@ -1,3 +1,4 @@
+from collections.abc import Callable
 from dataclasses import dataclass
 
 from quart import Quart
@@ -12,31 +13,42 @@ class RouteContext:
 
 
 class Route:
-    routes: list | dict
+    routes: dict[
+        str,
+        tuple[str, Callable]
+        | list[tuple[str, Callable]]
+        | tuple[str, Callable, str]
+        | list[tuple[str, Callable, str]],
+    ]
 
     def __init__(self, context: RouteContext):
         self.app = context.app
         self.config = context.config
 
     def register_routes(self):
-        def _add_rule(path, method, func):
+        def _add_rule(path, method, func, endpoint: str | None = None):
             # 统一添加 /api 前缀
             full_path = f"/api{path}"
-            self.app.add_url_rule(full_path, view_func=func, methods=[method])
+            self.app.add_url_rule(
+                full_path, view_func=func, methods=[method], endpoint=endpoint
+            )
 
-        # 兼容字典和列表两种格式
-        routes_to_register = (
-            self.routes.items() if isinstance(self.routes, dict) else self.routes
-        )
-
-        for route, definition in routes_to_register:
-            # 兼容一个路由多个方法
-            if isinstance(definition, list):
-                for method, func in definition:
-                    _add_rule(route, method, func)
+        for route, defi in self.routes.items():
+            if isinstance(defi, list):
+                for item in defi:
+                    if len(item) == 2:
+                        method, func = item
+                        _add_rule(route, method, func)
+                    elif len(item) == 3:
+                        method, func, endpoint = item
+                        _add_rule(route, method, func, endpoint)
             else:
-                method, func = definition
-                _add_rule(route, method, func)
+                if len(defi) == 2:
+                    method, func = defi
+                    _add_rule(route, method, func)
+                elif len(defi) == 3:
+                    method, func, endpoint = defi
+                    _add_rule(route, method, func, endpoint)
 
 
 @dataclass
