+++ astrbot/dashboard/routes/v1/README.md
@@ -1,0 +1,5 @@
+# AstrBot API
+
+为了更好地让外部系统与 AstrBot 进行交互，我们提供了可对外暴露的 API。
+
+并且 AstrBot WebUI 也将逐渐切换到此 API 规范来进行前后端交互。
