+++ astrbot/dashboard/routes/v1/__init__.py
@@ -1,0 +1,9 @@
+from .chat import V1ChatRoute
+from .knowledge_base import V1KnowledgeBaseRoute
+from .provider import V1ProviderRoute
+
+__all__ = [
+    "V1ChatRoute",
+    "V1KnowledgeBaseRoute",
+    "V1ProviderRoute",
+]
